Move utils.urlresolvers and utils management commands to EquiTrack
